requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1